\name{habR-package}
\alias{habR-package}
\alias{habR}
\docType{package}
\title{
\packageTitle{habR}
}
\description{
\packageDescription{habR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{habR}
\packageIndices{habR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{habR}

Maintainer: \packageMaintainer{habR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
